refactor: type 정의 파일 도메인 별로 분리
- '받은', '보낸' => 'received','sent'로 변경
