+++ frontend/src/components/Main/GridViewCoupon.tsx
@@ -1,4 +1,4 @@
-import { Coupon } from '../../types';
+import { Coupon } from '../../types/coupon';
 import CouponLayout from '../@shared/CouponLayout';
 
 const GridViewCoupon = ({ coupon }: { coupon: Coupon }) => {
