+++ frontend/src/components/SelectReceiver/CheckedUsers.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 import { BASE_URL } from '../../constants/api';
-import { UserProfile } from '../../types';
+import { UserProfile } from '../../types/user';
 
 const CheckedUsers = ({
   checkedUsers,
