+++ frontend/src/hooks/@queries/coupon.ts
@@ -1,13 +1,13 @@
-import { Axios, AxiosError } from 'axios';
-import { useEffect } from 'react';
-import { useMutation, useQuery, useQueryClient } from 'react-query';
+import { AxiosError } from 'axios';
+import { useMutation, useQuery } from 'react-query';
 import { client } from '../../apis/axios';
 import { API_PATH } from '../../constants/api';
-import { Coupon, CouponDetail, ErrorType } from '../../types';
+import { Coupon, CouponDetail } from '../../types/coupon';
+import { ErrorType } from '../../types/api';
 
 const SENT_OR_RECEIVED_API_PATH = {
-  받은: API_PATH.RECEIVED_COUPONS_ALL,
-  보낸: API_PATH.SENT_COUPONS,
+  received: API_PATH.RECEIVED_COUPONS_ALL,
+  sent: API_PATH.SENT_COUPONS,
 };
 
 export const COUPON_QUERY_KEY = {
@@ -15,7 +15,10 @@ export const COUPON_QUERY_KEY = {
   couponDetail: 'couponDetail',
 };
 
-export const useGetCouponDetail = (couponId, { onSuccess = () => {}, onError = () => {} } = {}) =>
+export const useGetCouponDetail = (
+  couponId: number,
+  { onSuccess = () => {}, onError = () => {} } = {}
+) =>
   useQuery<CouponDetail>(
     [COUPON_QUERY_KEY.couponDetail, couponId],
     () => getCouponDetailRequest(couponId),
