+++ frontend/src/hooks/EnterCouponContent/useEnterCouponContent.ts
@@ -3,7 +3,8 @@ import { useNavigate } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
 import { ROUTE_PATH } from '../../constants/routes';
 import { checkedUsersAtom } from '../../recoil/atom';
-import { CouponType, UserProfile } from '../../types';
+import { CouponType } from '../../types/coupon';
+import { UserProfile } from '../../types/user';
 import { usePostCouponMutation } from '../@queries/coupon';
 import { useGetUserProfile } from '../@queries/profile';
 import useModal from '../useModal';
