+++ frontend/src/pages/EnterCouponContent.tsx
@@ -5,15 +5,14 @@ import { Link } from 'react-router-dom';
 import ArrowBackButton from '../components/@shared/ArrowBackButton';
 import TabsNav from '../components/@shared/TabsNav';
 import useEnterCouponContent from '../hooks/EnterCouponContent/useEnterCouponContent';
-import { couponTypeKeys } from '../types';
 
 import CouponLayout from '../components/@shared/CouponLayout';
 import Header from '../layout/Header';
 import PageLayout from '../layout/PageLayout';
 import ConfirmCouponContentModal from '../components/EnterCouponContent/ConfirmCouponContentModal';
 import useModal from '../hooks/useModal';
-import { couponTypes } from '../types/index';
 import HeaderText from '../layout/HeaderText';
+import { couponTypeKeys, couponTypes } from '../types/coupon';
 
 const couponTypesWithoutEntire = couponTypeKeys.filter(type => type !== 'entire');
