+++ frontend/src/pages/Main.tsx
@@ -6,8 +6,6 @@ import GridViewCoupons from '../components/Main/GridViewCoupons';
 
 import { css } from '@emotion/react';
 import { Link } from 'react-router-dom';
-import { ROUTE_PATH } from '../constants/routes';
-import { couponTypeKeys, couponTypes } from '../types';
 import NoReceivedCoupon from './../components/@shared/noContent/NoReceivedCoupon';
 import NoSendCoupon from './../components/@shared/noContent/NoSendCoupon';
 import useMain from '../hooks/Main/useMain';
@@ -15,8 +13,9 @@ import HeaderText from '../layout/HeaderText';
 import MainPageLayout from '../layout/MainPageLayout';
 import useQRCoupon from '../hooks/useQRCoupon';
 import Spinner from '../components/@shared/Spinner';
+import { couponTypeKeys, couponTypes } from '../types/coupon';
 
-const sentOrReceivedArray = ['받은', '보낸'];
+const sentOrReceivedArray = ['received', 'sent'];
 
 const Main = () => {
   const {
@@ -40,17 +39,17 @@ const Main = () => {
         <S.SliderDiv length={2} current={sentOrReceivedArray.indexOf(sentOrReceived)} />
         <S.CouponStatusNav
           onClick={() => {
-            setSentOrReceived('받은');
+            setSentOrReceived('received');
           }}
-          selected={sentOrReceived === '받은'}
+          selected={sentOrReceived === 'received'}
         >
           <S.HeaderText>받은 쿠폰함</S.HeaderText>
         </S.CouponStatusNav>
         <S.CouponStatusNav
           onClick={() => {
-            setSentOrReceived('보낸');
+            setSentOrReceived('sent');
           }}
-          selected={sentOrReceived === '보낸'}
+          selected={sentOrReceived === 'sent'}
         >
           <S.HeaderText>보낸 쿠폰함</S.HeaderText>
         </S.CouponStatusNav>
@@ -81,7 +80,7 @@ const Main = () => {
           <Spinner />
         ) : coupons?.length ? (
           <GridViewCoupons coupons={coupons} />
-        ) : sentOrReceived === '보낸' ? (
+        ) : sentOrReceived === 'sent' ? (
           <NoSendCoupon />
         ) : (
           <NoReceivedCoupon />
