+++ frontend/src/recoil/atom.ts
@@ -1,6 +1,7 @@
 import { ReactNode } from 'react';
 import { atom } from 'recoil';
-import { CouponTransmitStatus, UserProfile } from '../types';
+import { CouponTransmitStatus } from '../types/coupon';
+import { UserProfile } from '../types/user';
 
 //checked users
 export const checkedUsersAtom = atom<UserProfile[]>({
@@ -32,7 +33,7 @@ export const targetCouponAtom = atom<any>({
 
 export const sentOrReceivedAtom = atom<CouponTransmitStatus>({
   key: 'sentOrReceived',
-  default: '받은',
+  default: 'received',
 });
 
 //toast
