+++ frontend/src/types/api.ts
@@ -1,0 +1,4 @@
+export interface ErrorType {
+  errorCode: string;
+  message: string;
+}
