+++ frontend/src/types/reservation.ts
@@ -1,0 +1,8 @@
+import { CouponStatus } from './coupon';
+import { MeetingTime } from './meeting';
+
+export type Reservation = {
+  reservationId: number;
+  status: CouponStatus;
+  time: MeetingTime | null;
+};
