+++ frontend/src/types/user.ts
@@ -1,0 +1,6 @@
+export interface UserProfile {
+  id: number;
+  name: string;
+  email: string;
+  imageUrl: string;
+}
